requirements.txt
flask==2.0.1
flask-socketio==5.1.1
python-dotenv==0.19.0
cryptography==3.4.7
PyJWT==2.1.0
numpy==1.21.2
eventlet==0.33.0
pytest==6.2.5
requests==2.26.0
python-engineio==4.2.1
gevent==21.8.0
gevent-websocket==0.10.1